vai que da

+++ Printf/ft_printf.h
@@ -6,28 +6,29 @@
 /*   By: adlopes- <adlopes-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/02 20:17:55 by adlopes-          #+#    #+#             */
-/*   Updated: 2023/08/18 23:26:27 by adlopes-         ###   ########.fr       */
+/*   Updated: 2023/09/11 19:40:39 by adlopes-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
+# include <stdint.h>
 # include <stdarg.h>
 # include <unistd.h>
 
 int		ft_printf(const char *format, ...);
 int		ft_print_char(char c);
 int		ft_print_str(char *str);
-int		ft_print_ptr(unsigned long ptr);
+int		ft_print_ptr(unsigned long long ptr);
 int		ft_print_nbr(int n);
 int		ft_print_unsig(unsigned int n);
 int		ft_print_hex(unsigned int num, const char format);
 int		ft_print_percent(void);
 
 void	ft_putstr(char *str);
-void	ft_put_ptr(unsigned long int num);
-int		ft_ptr_len(unsigned long int num);
+void	ft_put_ptr(unsigned long num);
+int		ft_ptr_len(unsigned long num);
 char	*ft_uitoa(unsigned int n);
 int		ft_num_len(unsigned	int num);
 void	ft_put_hex(unsigned int num, const char format);
